+++ src/Dolphin/OSAlloc.c
@@ -19,7 +19,7 @@ struct Heap* HeapArray;
 volatile OSHeapHandle __OSCurrHeap = -1;
 
 #define InRange(addr, start, end) ((u8*)(start) <= (u8*)(addr) && (u8*)(addr) < (u8*)(end))
-#define OFFSET(addr, align)       (((uintptr_t)(addr) & ((align)-1)))
+#define OFFSET(addr, align)       (((u32)(addr) & ((align)-1)))
 
 #define ALIGNMENT  32
 #define MINOBJSIZE 64
